+++ examples/hello_hardpy/test_simple.py
@@ -1,3 +1,2 @@
-
 def test_one():
     assert True